requirements.txt
distlib
wheel
setuptools
psycopg2-binary
asgiref>=3.3.2
autopep8==1.5
certifi==2023.7.22
cffi
chardet==3.0.4
cheroot==8.3.0
CherryPy==18.6.0
coverage==5.5
cryptography>=41.0.4
python-dateutil>=2.4.1
Django==4.2.11
django-cors-headers
django-expiring-token
django-permissions-policy
django-storages
django-rest-framework-proxy
django_rest_knox==4.2.0
django_storages
django_tenants
djangorestframework
fpdf==1.7.2
fpdf2==2.5.4
knox==0.1.14
more_itertools==8.2.0
numpy==1.24.1
protobuf==4.21.12
PyPDF2==3.0.1
python_dateutil==2.8.2
phpserialize==1.3
requests>=2.31.0
lxml>=4.5.1
pyotp>=2.8.0
portend==2.6
pycryptodome
schedule==0.6.0
setuptools==65.5.1
sphinx_rtd_theme==1.1.1
six==1.14.0
sqlparse>=0.4.4
stripe>=5.0.0
tempora==3.0.0
tqdm==4.66.3
typing-extensions==3.10.0.2
urllib3==1.26.18
whitenoise>=6.2.0
zc.lockfile==2.0
zipp==3.5.0
google-cloud-storage
django-storages[google]
google-auth
django-debug-toolbar 
django-storages[azure] 
trimesh 
networkx 
django-cryptography
djangorestframework-api-key==3.0.0